% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{custom_poly_to_sf}
\alias{custom_poly_to_sf}
\title{sf builder for leaflet custom polygons plugin}
\usage{
custom_poly_to_sf(custom_polygon)
}
\arguments{
\item{custom_polygon}{object as the one returned by the leaflet plugin}
}
\description{
Helper to create sf objects from leaflet polygons
}
