% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{custom_polygon_filter_expr}
\alias{custom_polygon_filter_expr}
\title{Custom polygon filter expression builder}
\usage{
custom_polygon_filter_expr(custom_polygon, conn)
}
\arguments{
\item{custom_polygon}{sf or sp object with the custom polygon to filter plots by}

\item{conn}{pool object with the database connection}
}
\description{
Custom polygon filter expression builder
}
