% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{nfi_close}
\alias{nfi_close}
\title{Closing connection to the NFI SQL database}
\usage{
nfi_close(conn)
}
\arguments{
\item{conn}{}
}
\description{
Wrapper of \code{\link{poolClose}} to close the connection
}
