% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{nfi_connect}
\alias{nfi_connect}
\title{Creating a connection to the NFI SQL database}
\usage{
nfi_connect(user = "guest", password = "guest", dbname = "tururu",
  idleTimeout = 3600000)
}
\arguments{
\item{user}{character with database user, default to "guest"}

\item{password}{character woth the user password, default to "guest"}

\item{dbname}{character with the database address}

\item{idleTimeout}{Idle time in seconds before the connection expires}
}
\description{
Wrapper of \code{\link{dbPool}} to connect to the NFI database
}
\details{
By default (no arguments supplied) it allows to access as a guest to the
database
}
