% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{nfi_results_data}
\alias{nfi_results_data}
\title{Retrieve NFI data as tbl}
\usage{
nfi_results_data(conn, nfi = c("nfi_2", "nfi_3", "nfi_4", "nfi_2_nfi_3",
  "nfi_3_nfi_4", "nfi_2_shrub", "nfi_3_shrub", "nfi_4_shrub",
  "nfi_2_regen", "nfi_3_regen", "nfi_4_regen"),
  functional_group = "none", diameter_classes = FALSE,
  .collect = TRUE)
}
\arguments{
\item{conn}{pool object to access the tables, as obtained from
\code{\link{nfi_connect}}}

\item{nfi}{character indicating the nfi version}

\item{functional_group}{Functional group to retrieve table for,
Default to 'none' (no functional group). See details for more information}

\item{diameter_classes}{logical indicating if diameter classes are required
(breaking down the table by diameter classes). Default to FALSE.}

\item{.collect}{Logical indicating if the tbl must be collected locally.
Default to TRUE}
}
\description{
Get the results data by plot (with or without breakdown)
}
\details{
Data is retrieved from the postgreSQL db in tbl form
}
\section{Functional groups}{

\code{functional_group} parameter allows to retrieve the table of plots
  broken down by the desired group. Allowed values are:
  \itemize{
    \item{\code{"none"} (No breakdown)}
    \item{\code{"species"}}
    \item{\code{"simpspecies"}}
    \item{\code{"genus"}}
    \item{\code{"dec"} (Deciduous/Esclerophyllous/Conifer)}
    \item{\code{"bc"} (Broadleaf/Conifer)}
  }
}

