% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modifications.R
\name{nfi_results_filter}
\alias{nfi_results_filter}
\title{Filter the nfi_results table}
\usage{
nfi_results_filter(nfi_data, variables, conn, ..., .collect = TRUE)
}
\arguments{
\item{nfi_data}{tbl from \code{\link{nfi_results_data}}}

\item{variables}{character vector with the variables to filter by}

\item{conn}{pool connection to the database}

\item{...}{Filter expressions, in the same order as \code{variables} argument}

\item{.collect}{Logical indicating if the tbl must be collected locally.
Default to TRUE}
}
\description{
Filter the results table by vars in the PLOTS and PLOTS_DYNAMIC tables as
well as by columns in the data itself
}
