% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modifications.R
\name{nfi_results_summarise}
\alias{nfi_results_summarise}
\title{Summarise the nfi results data (raw or filtered)}
\usage{
nfi_results_summarise(nfi_data, polygon_group, functional_group = "none",
  diameter_classes, .funs = dplyr::funs(mean = mean(., na.rm = TRUE), sd =
  sd(., na.rm = TRUE), min = min(., na.rm = TRUE), max = max(., na.rm = TRUE)),
  .collect = TRUE)
}
\arguments{
\item{nfi_data}{tbl from \code{\link{nfi_results_data}} or
\code{\link{nfi_results_filter}}}

\item{polygon_group}{character indicating the polygon group to summarise}

\item{functional_group}{Functional group to retrieve table for. Default to 'none'
(no functional group). See details for more information}

\item{diameter_classes}{Logical indicating if the table contains diameter classes}

\item{.funs}{functions list (as obtained from \code{\link[dplyr]{funs}}) with the
summarise functions}

\item{.collect}{Logical indicating if the tbl must be collected locally.
Default to TRUE}
}
\description{
Summarise the data by polygon groups and functional groups (if any)
}
\details{
turururururur
}
\section{Functional groups}{

\code{functional_group} parameter allows to retrieve the table of plots
  broken down by the desired group. Allowed values are:
  \itemize{
    \item{\code{"none"} (No breakdown)}
    \item{\code{"species"}}
    \item{\code{"simplified_species"}}
    \item{\code{"genus"}}
    \item{\code{"dec"} (Deciduous/Esclerophyllous/Conifer)}
    \item{\code{"bc"} (Broadleaf/Conifer)}
  }
}

