% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nif_get_scenario}
\alias{nif_get_scenario}
\title{NFI scenario}
\usage{
nif_get_scenario(viz_shape, agg_level)
}
\arguments{
\item{viz_shape}{The shape to visualize, plots or polygons}

\item{agg_level}{The breakdown level, none, species...}
}
\description{
Return the current scenario as character
}
\details{
There are four main scenarios to work with in the NFI data:
\itemize{
  \item{Scenario 1. Total by plot. In this scenario we can have the raw total
  and the dominant functional group (in two flavours, basal area or density)}
  \item{Scenario 2. Plot broken down by functional group, regardless the
  dominance}
  \item{Scenarios 3 and 4. The same as 1 and 2 but summarised by polygon,
  being the polygon the administrative divisions or custom polygons}
}
}
\section{viz_shape}{

\code{plot} or \code{polygon}
}

\section{agg_level}{

\code{none}, \code{species}, \code{simpspecies}, \code{genus}, \code{dec} or \code{bc}
}

