% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{var_in_PLOTS_DYNAMIC}
\alias{var_in_PLOTS_DYNAMIC}
\title{var_in_PLOTS_DYNAMIC}
\usage{
var_in_PLOTS_DYNAMIC(variables, nfi, conn)
}
\description{
Given a variables character list, returns an index with which vars are in the
corresponding PLOTS_*_DYNAMIC_INFO table
}
