% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vars_in_nfi_data}
\alias{vars_in_nfi_data}
\title{vars_in_nfi_data}
\usage{
vars_in_nfi_data(variables, nfi_data)
}
\arguments{
\item{variables}{character vector with the names of the variables to check}

\item{nfi_data}{nfi_data}
}
\description{
helper to get the names in data, checking if it is already collected or not
}
