% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vars_in_PLOTS}
\alias{vars_in_PLOTS}
\title{vars_in_PLOTS}
\usage{
vars_in_PLOTS(variables, conn)
}
\description{
Given a variables character list, returns an index with which vars are in the
PLOTS table
}
